+++ plone/recipe/codeanalysis/imports.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from plone.recipe.codeanalysis.analyser import console_factory
 from plone.recipe.codeanalysis.clean_lines import CleanLines
 import re
 
@@ -78,4 +79,4 @@ class Imports(CleanLines):
 
 
 def console_script(options):
-    return Imports(options).run()
+    console_factory(Imports, options)