+++ plone/recipe/codeanalysis/jshint.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 from plone.recipe.codeanalysis.analyser import Analyser
+from plone.recipe.codeanalysis.analyser import console_factory
 import re
 
 
@@ -43,4 +44,4 @@ class JSHint(Analyser):
 
 
 def console_script(options):
-    return JSHint(options).run()
+    console_factory(JSHint, options)